Start refactor create proposal
