+++ src/components/SpaceCreateContent.vue
@@ -1,0 +1,147 @@
+<script setup lang="ts">
+import { ref } from 'vue';
+import { useIntl } from '@/composables/useIntl';
+import { useImageUpload } from '@/composables/useImageUpload';
+import { useSpaceCreateStore } from '@/composables/useSpaceCreateStore';
+
+defineProps<{
+  preview: boolean;
+  bodyLimit: number;
+}>();
+
+const { formatNumber } = useIntl();
+const { form } = useSpaceCreateStore();
+
+const imageDragging = ref(false);
+const textAreaEl = ref<HTMLTextAreaElement | null>(null);
+
+const injectImageToBody = image => {
+  const cursorPosition = textAreaEl.value?.selectionStart;
+  const currentBody = textAreaEl.value?.value;
+  const currentBodyWithImage = `${currentBody?.substring(
+    0,
+    cursorPosition
+  )} \n![${image.name}](${image.url})
+    ${currentBody?.substring(cursorPosition as number)}`;
+
+  form.value.body = currentBodyWithImage;
+};
+
+const {
+  upload,
+  error: imageUploadError,
+  uploading
+} = useImageUpload({
+  onSuccess: injectImageToBody
+});
+
+const handlePaste = e => {
+  for (let i = 0; i < e.clipboardData.items.length; ++i) {
+    let item = e.clipboardData.items[i];
+    if (item.kind == 'file' && item.type.startsWith('image/')) {
+      const file = item.getAsFile();
+      upload(new File([file], 'image', { type: file.type }));
+    }
+  }
+};
+
+const handleDrop = e => {
+  for (let i = 0; i < e.dataTransfer.files.length; i++) {
+    let item = e.dataTransfer.files[i];
+    if (item.type.startsWith('image/')) {
+      upload(item);
+    }
+  }
+};
+</script>
+
+<template>
+  <div class="px-4 md:px-0">
+    <div class="flex flex-col space-y-3">
+      <h1
+        v-if="preview"
+        v-text="form.name || $t('create.untitled')"
+        class="w-full break-all"
+      />
+      <SBaseInput
+        v-else
+        v-model="form.name"
+        :title="$t('create.proposalTitle')"
+        :maxLength="128"
+        focusOnMount
+      />
+
+      <div v-if="!preview">
+        <div class="flex justify-between">
+          <SBaseLabel>
+            {{ $t('create.proposalDescription') }}
+          </SBaseLabel>
+          <div class="text-xs">
+            {{ formatNumber(form.body.length) }} /
+            {{ formatNumber(bodyLimit) }}
+          </div>
+        </div>
+        <div
+          @drop.prevent="handleDrop"
+          @dragover="imageDragging = true"
+          @dragleave="imageDragging = false"
+        >
+          <div
+            class="min-h-[240px] peer border rounded-t-xl overflow-hidden focus-within:border-skin-text"
+          >
+            <textarea
+              @paste="handlePaste"
+              ref="textAreaEl"
+              class="s-input pt-0 w-full min-h-[240px] border-none !rounded-xl text-base h-full mt-0"
+              :maxLength="bodyLimit"
+              v-model="form.body"
+            />
+          </div>
+
+          <label
+            class="relative flex justify-between border border-skin-border rounded-b-xl py-1 px-2 items-center peer-focus-within:border-skin-text border-t-0"
+          >
+            <input
+              accept="image/jpg, image/jpeg, image/png"
+              type="file"
+              class="opacity-0 absolute p-[5px] top-0 right-0 bottom-0 left-0 w-full ml-0"
+              @change="e => upload((e.target as HTMLInputElement)?.files?.[0])"
+            />
+
+            <span class="pointer-events-none relative pl-1 text-sm">
+              <span v-if="uploading" class="flex">
+                <LoadingSpinner small class="mr-2 -mt-[2px]" />
+                {{ $t('create.uploading') }}
+              </span>
+              <span v-else-if="imageUploadError !== ''">
+                {{ imageUploadError }}</span
+              >
+              <span v-else>
+                {{ $t('create.uploadImageExplainer') }}
+              </span>
+            </span>
+            <BaseLink
+              class="relative inline"
+              link="https://docs.github.com/github/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax"
+              v-tippy="{ content: $t('create.markdown') }"
+              hide-external-icon
+            >
+              <BaseIcon name="markdown" class="text-skin-text" />
+            </BaseLink>
+          </label>
+        </div>
+      </div>
+
+      <div v-if="form.body && preview" class="mb-4">
+        <BaseMarkdown :body="form.body" />
+      </div>
+
+      <SBaseInput
+        v-if="!preview"
+        v-model.trim="form.discussion"
+        placeholder="e.g. https://forum.balancer.fi/proposal..."
+        :title="$t('create.discussion')"
+      />
+    </div>
+  </div>
+</template>
